+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 2.0.11
 - [SRN-15136 Meus Endereços - Criar Endereço](https://dev.azure.com/ARECO/Canal%20App%20e%20Omni/_workitems/edit/15136)
-
+- [SRN-15169 Evento view_cart](https://dev.azure.com/ARECO/Canal%20App%20e%20Omni/_workitems/edit/15169)
 ---
 
 ### 2.0.10